fix: correct Service.price type, enforce SQLite foreign keys (2.4.2, 2.4.3)
2.4.2: Service.price is a Numeric(10,2) column, which SQLAlchemy
returns as Decimal, not float - the Mapped[Optional[float]] annotation
was wrong (confirmed no code anywhere reads .price yet, so no runtime
fallout, but it was a landmine for whoever adds pricing display next).
Fixed here and in the ServiceCreateData/ServiceUpdateData DTOs.

2.4.3: SQLite doesn't enforce foreign keys unless explicitly told to,
unlike PostgreSQL (used in production). Without it, Booking's
ondelete=CASCADE/RESTRICT/SET NULL was silently a no-op under SQLite
(dev/tests), so an integrity bug there could reach production
undetected. Adds enable_sqlite_foreign_keys(), applied to the app's
engine when running on SQLite.

+++ backend/app/dto.py
@@ -13,6 +13,7 @@ import from the other's package.
 """
 
 from dataclasses import dataclass
+from decimal import Decimal
 from typing import Optional
 
 
@@ -22,7 +23,7 @@ class ServiceCreateData:
     name_pl: str
     name_ru: str
     duration_minutes: int
-    price: Optional[float] = None
+    price: Optional[Decimal] = None
     description_pl: Optional[str] = None
     description_ru: Optional[str] = None
 
@@ -35,6 +36,6 @@ class ServiceUpdateData:
     name_pl: Optional[str] = None
     name_ru: Optional[str] = None
     duration_minutes: Optional[int] = None
-    price: Optional[float] = None
+    price: Optional[Decimal] = None
     description_pl: Optional[str] = None
     description_ru: Optional[str] = None
